Fix PHP 8.4 SecurityBundle PHPUnit grind (365 tests, exit 0)
Align AuthenticatorManager DI with Http component constructor (remove stale
erase_credentials argument). Harden strict typing across Security, RateLimiter,
CSRF, LoginLink, RememberMe, ExpressionLanguage, and ErrorHandler paths used by
SecurityBundle functional tests. Adjust test harness for Reference/ClassStub
casts, readonly InMemoryUser, and vendor JWT deprecations.

Add build/junit-security-bundle-20260913.xml from filtered grind.

+++ src/Symfony/Component/Security/Csrf/CsrfTokenManager.php
@@ -130,7 +130,7 @@ class CsrfTokenManager implements CsrfTokenManagerInterface
     private function xor(string $value, string $key): string
     {
         if (\strlen($value) > \strlen($key)) {
-            $key = str_repeat($key, ceil(\strlen($value) / \strlen($key)));
+            $key = str_repeat($key, (int) ceil(\strlen($value) / \strlen($key)));
         }
 
         return $value ^ $key;
